Refactor: #118 ChatRoom, UserChatRoom 삭제,

+++ cookie/src/main/java/com/cookie/admin/dto/response/AdminBadges.java
@@ -11,5 +11,5 @@ public class AdminBadges {
     private String badgeName;
     private String genre;
     private String badgeImage;
-    private Integer needPoint;
+    private Long needPoint;
 }
